@@ -29,16 +29,17 @@
   },
   "dependencies": {
     "@nestjs-modules/mailer": "~1.6.0",
-    "@nestjs/common": "~7.6.18",
-    "@nestjs/core": "~7.6.18",
-    "@nestjs/jwt": "~7.2.0",
-    "@nestjs/passport": "~7.1.5",
-    "@nestjs/platform-express": "~7.6.18",
-    "@nestjs/platform-socket.io": "~7.6.18",
-    "@nestjs/serve-static": "~2.1.4",
-    "@nestjs/swagger": "~4.8.1",
-    "@nestjs/typeorm": "~7.1.5",
-    "@nestjs/websockets": "~7.6.18",
+    "@nestjs/axios": "^0.0.1",
+    "@nestjs/common": "~8.0.0",
+    "@nestjs/core": "~8.0.0",
+    "@nestjs/jwt": "~8.0.0",
+    "@nestjs/passport": "~7.1.6",
+    "@nestjs/platform-express": "~8.0.0",
+    "@nestjs/platform-socket.io": "~8.0.0",
+    "@nestjs/serve-static": "~2.2.2",
+    "@nestjs/swagger": "~4.8.2",
+    "@nestjs/typeorm": "~8.0.0",
+    "@nestjs/websockets": "~8.0.0",
     "bcryptjs": "~2.4.3",
     "camelcase": "~6.2.0",
     "class-transformer": "~0.4.0",
@@ -61,7 +62,7 @@
     "pg-error-enum": "~0.5.0",
     "reflect-metadata": "~0.1.13",
     "rimraf": "~3.0.2",
-    "rxjs": "~7.1.0",
+    "rxjs": "~7.2.0",
     "st-utils": "~1.8.0",
     "swagger-ui-express": "~4.1.6",
     "type-fest": "~1.2.1",
@@ -71,41 +72,40 @@
     "uuid": "~8.3.2"
   },
   "devDependencies": {
-    "@nestjs/cli": "~7.6.0",
-    "@nestjs/schematics": "~7.3.1",
-    "@nestjs/testing": "~7.6.18",
+    "@nestjs/cli": "~8.0.2",
+    "@nestjs/schematics": "~8.0.0",
+    "@nestjs/testing": "~8.0.0",
     "@types/adm-zip": "~0.4.34",
     "@types/bcryptjs": "~2.4.2",
-    "@types/compression": "~1.7.0",
-    "@types/config": "~0.0.38",
-    "@types/express": "~4.17.12",
-    "@types/fs-extra": "~9.0.11",
+    "@types/compression": "~1.7.1",
+    "@types/config": "~0.0.39",
+    "@types/express": "~4.17.13",
+    "@types/fs-extra": "~9.0.12",
     "@types/inline-css": "~3.0.0",
-    "@types/jest": "~26.0.23",
-    "@types/morgan": "~1.9.2",
+    "@types/jest": "~26.0.24",
+    "@types/morgan": "~1.9.3",
     "@types/node": "~14.14.45",
     "@types/nodemailer": "~6.4.2",
     "@types/openid": "~2.0.1",
     "@types/passport-jwt": "~3.0.5",
-    "@types/socket.io": "~2.1.13",
     "@types/supertest": "~2.0.11",
-    "@types/uuid": "~8.3.0",
-    "@typescript-eslint/eslint-plugin": "~4.28.0",
-    "@typescript-eslint/parser": "~4.28.0",
+    "@types/uuid": "~8.3.1",
+    "@typescript-eslint/eslint-plugin": "~4.28.2",
+    "@typescript-eslint/parser": "~4.28.2",
     "adm-zip": "~0.5.5",
-    "eslint": "~7.29.0",
+    "eslint": "~7.30.0",
     "eslint-config-prettier": "~8.3.0",
     "eslint-plugin-import": "~2.23.4",
     "fs-extra": "~10.0.0",
-    "jest": "~26.6.3",
+    "jest": "~27.0.6",
     "ora": "~5.4.1",
     "prettier": "~2.3.2",
     "supertest": "~6.1.3",
-    "ts-jest": "~26.5.6",
-    "ts-loader": "~8.1.0",
-    "ts-node": "~9.1.1",
-    "tsconfig-paths": "~3.9.0",
-    "typescript": "~4.3.4",
+    "ts-jest": "~27.0.3",
+    "ts-loader": "~9.2.3",
+    "ts-node": "~10.0.0",
+    "tsconfig-paths": "~3.10.1",
+    "typescript": "~4.3.5",
     "yargs": "~17.0.1"
   },
   "jest": {
