@@ -3,8 +3,7 @@ import { MailerService } from '@nestjs-modules/mailer';
 import { MailInfoTemplate } from './mail-info.interface';
 import { environment } from '../environment/environment';
 import { MailQueueRepository } from './mail-queue.repository';
-import { of, Subject } from 'rxjs';
-import { auditTime, catchError, filter, switchMap } from 'rxjs/operators';
+import { auditTime, catchError, filter, of, Subject, switchMap } from 'rxjs';
 import { Transactional } from 'typeorm-transactional-cls-hooked';
 import { MailQueue } from './mail-queue.entity';
 import { MailPriorityEnum } from './mail-priority.enum';
